comments and reorganize
